ambiR: Calculate AZTI’s Marine Biotic Index

Calculate AZTI’s Marine Biotic Index - AMBI. The included list of benthic fauna species according to their sensitivity to pollution. Matching species in sample data to the list allows the calculation of fractions of individuals in the different sensitivity categories and thereafter the AMBI index. The Shannon Diversity Index H' and the Danish benthic fauna quality index DKI (Dansk Kvalitetsindeks) can also be calculated, as well as the multivariate M-AMBI index. Borja, A., Franco, J. ,Pérez, V. (2000) "A marine biotic index to establish the ecological quality of soft bottom benthos within European estuarine and coastal environments" doi:10.1016/S0025-326X(00)00061-8 .

See also

Author

Maintainer: Ciarán J. Murray cjm@niva-dk.dk (ORCID) [copyright holder]

Authors:

Ángel Borja aborja@azti.es (ORCID)

Sarai Pouso spouso@azti.es (ORCID)

Iñigo Muxika imuxika@azti.es (ORCID)

Joxe Mikel Garmendia jgarmendia@azti.es (ORCID)

Other contributors:

Steen Knudsen steen.knudsen@niva-dk.dk (ORCID) [contributor]

GES4SEAS (Grant Agreement 101059877 - GES4SEAS. The GES4SEAS project has been approved under the HORIZON-CL6-2021-BIODIV-01-04 call: 'Assess and predict integrated impacts of cumulative direct and indirect stressors on coastal and marine biodiversity, ecosystems and their services'. Funded by the European Union. Views and opinions expressed are however those of the authors only and do not necessarily reflect those of the European Union or UK Research and Innovation. Neither the European Union nor the granting authority can be held responsible for them.) [funder]