Calculates DKI (v1)

Calculates the Danish quality index DKI (v1)

The DKI is based on AMBI and can only be calculated after first calculating AMBI, the AZTI Marine Biotic Index, and H', the Shannon diversity index. Both indices are included in output from the function AMBI().

The function uses an estimated maximum value of H' H_max in Danish waters as a reference value to normalise DKI. If this value is not specified as an argument, the function uses the default value (5.0)

Usage

DKI(AMBI, H, N, S, H_max = 5)

Source

The Danish quality index is described in:

Carstensen, J., Krause-Jensen, D., Josefson, A. (2014) Development and testing of tools for intercalibration of phytoplankton, macrovegetation and benthic fauna in Danish coastal areas. Scientific Report from DCE – Danish Centre for Environment and Energy No. 93. https://dce2.au.dk/pub/SR93.pdf

Reference for default value of H_max = 5:

"However, in the present exercise, the Danish method used Hmax (∼5) as a kind of reference"

Borja, A., Josefson, A.B., Miles, A., Muxika, I., Olsgaard, F., Phillips, G., Rodríguez, J.G. & Rygg, B. 2007: An approach to the intercalibration of benthic ecological status assessment in the North Atlantic ecoregion, according to the European Water Framework Directive. - Marine Pollution Bulletin 55:42-52. https://doi.org/10.1016/j.marpolbul.2006.08.018

Arguments

AMBI

AMBI, the AZTI Marine Biotic Index, calculated using AMBI()

H

H', the Shannon diversity index, calculated using Hdash()

N

number of individuals - generated by both AMBI() and Hdash()

S

number of species - generated by both AMBI() and Hdash()

H_max

maximum H' used to normalise AMBI, default 5

Value

DKI index value

Details

The AMBI() and Hdash() functions take a dataframe of observations as an argument. The DKI functions, DKI2() and DKI(), do not take a dataframe as an argument. Instead they take values of the input parameters, either single values or as vectors.

To calculate DKI for a dataframe of AMBI values, it could be called from e.g. within a dplyr::mutate() function call. See the examples below.

See also

Examples

# Simple example

DKI(AMBI = 1.61, H = 2.36, N = 25, S = 6)
#> [1] 0.55683

# ------ Example workflow for calculating DKI from species counts ----

# calculate AMBI index
dfAMBI <- AMBI(test_data, by = c("station"), var_rep="replicate")[["AMBI"]]

# show AMBI results
dfAMBI
#> # A tibble: 3 × 12
#>   station  AMBI     H     S   fNA     N     I    II   III     IV      V
#>     <dbl> <dbl> <dbl> <int> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1       1  1.48  1.80     6     0    16 0.125 0.75  0.125 0      0     
#> 2       2  1.89  3.54    22     0    80 0.4   0.138 0.3   0.15   0.0125
#> 3       3  4.12  2.50     9     0    24 0     0.125 0.292 0.0833 0.5   
#> # ℹ 1 more variable: Disturbance <chr>

# calculate DKI from AMBI results
dplyr::mutate(dfAMBI, DKI = DKI(AMBI, H, N, S))
#> # A tibble: 3 × 13
#>   station  AMBI     H     S   fNA     N     I    II   III     IV      V
#>     <dbl> <dbl> <dbl> <int> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1       1  1.48  1.80     6     0    16 0.125 0.75  0.125 0      0     
#> 2       2  1.89  3.54    22     0    80 0.4   0.138 0.3   0.15   0.0125
#> 3       3  4.12  2.50     9     0    24 0     0.125 0.292 0.0833 0.5   
#> # ℹ 2 more variables: Disturbance <chr>, DKI <dbl>